router tests done and passed
